Move EntityTable into correct package
