+++ caps-spark/src/main/scala/org/opencypher/caps/api/schema/Entity.scala
@@ -21,6 +21,10 @@ case class Labels(labels: String*) extends StaticAnnotation
 
 case class RelationshipType(relType: String) extends StaticAnnotation
 
+object Entity {
+  private[schema] val sourceIdKey = "id"
+}
+
 sealed trait Entity extends Product {
   def id: Long
 }
@@ -31,6 +35,15 @@ sealed trait Entity extends Product {
   * then the labels from that annotation are used instead.
   */
 trait Node extends Entity
+
+object Relationship {
+  private[schema] val sourceStartNodeKey = "source"
+
+  private[schema] val sourceEndNodeKey = "target"
+
+  private[schema] val nonPropertyAttributes =
+    Set(Entity.sourceIdKey, sourceStartNodeKey, sourceEndNodeKey)
+}
 
 /**
   * If a relationship has no type annotation, then the class name in all caps is used as its type.
