+++ caps-spark/src/main/scala/org/opencypher/caps/impl/spark/CAPSGraph.scala
@@ -16,11 +16,11 @@
 package org.opencypher.caps.impl.spark
 
 import org.apache.spark.storage.StorageLevel
+import org.opencypher.caps.api.CAPSSession
 import org.opencypher.caps.api.exception.IllegalArgumentException
 import org.opencypher.caps.api.graph.PropertyGraph
-import org.opencypher.caps.api.schema.Schema
+import org.opencypher.caps.api.schema.{EntityTable, NodeTable, Schema}
 import org.opencypher.caps.api.types.{CTNode, CTRelationship}
-import org.opencypher.caps.api.{CAPSSession, EntityTable, NodeTable}
 import org.opencypher.caps.impl.record.{OpaqueField, RecordHeader, _}
 import org.opencypher.caps.impl.spark.CAPSConverters._
 import org.opencypher.caps.ir.api.expr._
