+++ caps-spark/src/main/scala/org/opencypher/caps/impl/spark/CAPSRecords.scala
@@ -23,11 +23,12 @@ import org.apache.spark.sql._
 import org.apache.spark.sql.catalyst.expressions.GenericRowWithSchema
 import org.apache.spark.sql.types._
 import org.apache.spark.storage.StorageLevel
+import org.opencypher.caps.api.CAPSSession
 import org.opencypher.caps.api.exception.{DuplicateSourceColumnException, IllegalArgumentException, IllegalStateException}
 import org.opencypher.caps.api.io.conversion.{NodeMapping, RelationshipMapping}
+import org.opencypher.caps.api.schema.{EntityTable, NodeTable, RelationshipTable}
 import org.opencypher.caps.api.types._
 import org.opencypher.caps.api.value._
-import org.opencypher.caps.api.{CAPSSession, EntityTable, NodeTable, RelationshipTable}
 import org.opencypher.caps.impl.record.CAPSRecordHeader._
 import org.opencypher.caps.impl.record.{CAPSRecordHeader, _}
 import org.opencypher.caps.impl.spark.CAPSRecords.{prepareDataFrame, verifyAndCreate}
