+++ caps-spark/src/test/scala/org/opencypher/caps/impl/record/EntityTableTest.scala
@@ -17,10 +17,9 @@ package org.opencypher.caps.impl.record
 
 import org.apache.spark.sql.Row
 import org.opencypher.caps.api.io.conversion.{NodeMapping, RelationshipMapping}
-import org.opencypher.caps.api.schema.Schema
+import org.opencypher.caps.api.schema.{NodeTable, RelationshipTable, Schema}
 import org.opencypher.caps.api.types.{CTFloat, CTInteger, CTString}
 import org.opencypher.caps.api.value.CAPSMap
-import org.opencypher.caps.api.{NodeTable, RelationshipTable}
 import org.opencypher.caps.demo.SocialNetworkData.{Friend, Person}
 import org.opencypher.caps.impl.spark.CAPSGraph
 import org.opencypher.caps.test.CAPSTestSuite
